requirements.txt
fastapi
uvicorn[standard]
requests
numpy
playwright
redis
celery
chromadb
sentence-transformers
python-dotenv 